+++ meridian/__init__.py
@@ -14,7 +14,7 @@
 
 """Meridian API."""
 
-__version__ = "1.0.3"
+__version__ = "1.0.4"
 
 
 from meridian import analysis